Separate concurrency function
Pass by reference for calc_pyramid_width with pyramid

Now separate function for concurrent and nonconcurrent functions. Can't recall if c uses concurrency but it should be easy to port to a concurrent c++ framework.

Maybe refactor the printing functions to accept a function that prints and uses that to print characters.